system settings: move system settings to samples
System settings (for runtime PSK setting) is moved to samples/common
