+++ net/golioth/Kconfig
@@ -293,17 +293,6 @@ config GOLIOTH_SYSTEM_CLIENT_RX_BUF_SIZE
 	  Size of receive buffer, which is used for reading data from network
 	  socket.
 
-config GOLIOTH_SYSTEM_SETTINGS
-	bool "Load credentials from persistent settings"
-	default y
-	depends on GOLIOTH_AUTH_METHOD_PSK
-	depends on SETTINGS
-	help
-	  When selected, Golioth credentials will be loaded from settings
-	  subsystem.
-
-if GOLIOTH_SYSTEM_SETTINGS
-
 config GOLIOTH_SYSTEM_CLIENT_PSK_ID_MAX_LEN
 	int "Max length of PSK ID"
 	default 64
@@ -312,12 +301,10 @@ config GOLIOTH_SYSTEM_CLIENT_PSK_ID_MAX_LEN
 
 config GOLIOTH_SYSTEM_CLIENT_PSK_MAX_LEN
 	int "Max length of PSK"
-	default MBEDTLS_PSK_MAX_LEN if MBEDTLS_BUILTIN
+	default MBEDTLS_PSK_MAX_LEN if (MBEDTLS_BUILTIN && GOLIOTH_AUTH_METHOD_PSK)
 	default 64
 	help
 	  Maximum length of PSK, in bytes.
-
-endif # GOLIOTH_SYSTEM_SETTINGS
 
 endif # GOLIOTH_SYSTEM_CLIENT
