+++ samples/test/prj.conf
@@ -15,6 +15,7 @@ CONFIG_GOLIOTH_SAMPLE_HARDCODED_CREDENTIALS=n
 
 CONFIG_SETTINGS=y
 CONFIG_SETTINGS_RUNTIME=y
+CONFIG_GOLIOTH_SAMPLE_PSK_SETTINGS=y
 CONFIG_GOLIOTH_SAMPLE_SETTINGS_AUTOLOAD=y
 CONFIG_GOLIOTH_SAMPLE_SETTINGS_SHELL=y
